core/internal/clientstate: Split tests into separate files
Reorganize test to separate files according to the area of
functionality tested.
