+++ core/internal/clientstate/request-seq_test.go
@@ -1,0 +1,257 @@
+// Copyright (c) 2018 NEC Laboratories Europe GmbH.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package clientstate
+
+import (
+	"fmt"
+	"sync"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestReqeustSeq(t *testing.T) {
+	t.Run("CaptureRelease", testCaptureReleaseRequestSeq)
+	t.Run("CaptureReleaseConcurrent", testCaptureReleaseRequestSeqConcurrent)
+	t.Run("Prepare", testPrepareRequestSeq)
+	t.Run("Retire", testRetireRequestSeq)
+}
+
+func testCaptureReleaseRequestSeq(t *testing.T) {
+	s := New()
+
+	cases := []struct {
+		desc string
+		seq  int
+
+		capture bool
+		release bool
+
+		ok  bool
+		new bool
+	}{{
+		desc:    "Capture and release the first ID",
+		seq:     100,
+		capture: true,
+		release: true,
+		ok:      true,
+		new:     true,
+	}, {
+		desc:    "Release the same ID again",
+		seq:     100,
+		release: true,
+		ok:      false,
+	}, {
+		desc:    "Release new ID before capturing",
+		seq:     200,
+		release: true,
+		ok:      false,
+	}, {
+		desc:    "Capture another ID",
+		seq:     200,
+		capture: true,
+		ok:      true,
+		new:     true,
+	}, {
+		desc:    "Capture older ID",
+		seq:     50,
+		capture: true,
+		new:     false,
+	}, {
+		desc:    "Release not captured ID",
+		seq:     150,
+		release: true,
+		ok:      false,
+	}, {
+		desc:    "Release last captured ID",
+		seq:     200,
+		release: true,
+		ok:      true,
+	}}
+
+	for _, c := range cases {
+		seq := uint64(c.seq)
+		if c.capture {
+			new := s.CaptureRequestSeq(seq)
+			require.Equal(t, c.new, new, c.desc)
+		}
+		if c.release {
+			err := s.ReleaseRequestSeq(seq)
+			if c.ok {
+				require.NoError(t, err, c.desc)
+			} else {
+				require.Error(t, err, c.desc)
+			}
+		}
+	}
+}
+
+func testCaptureReleaseRequestSeqConcurrent(t *testing.T) {
+	const nrConcurrent = 5
+	const nrSeqs = 10
+
+	seqs := make([]bool, nrSeqs)
+	wg := new(sync.WaitGroup)
+
+	state := New()
+
+	wg.Add(nrConcurrent)
+	for workerID := 0; workerID < nrConcurrent; workerID++ {
+		workerID := workerID
+
+		go func() {
+			defer wg.Done()
+
+			for seq := 1; seq <= nrSeqs; seq++ {
+				assertMsg := fmt.Sprintf("Worker %d, seq %d", workerID, seq)
+
+				if new := state.CaptureRequestSeq(uint64(seq)); new {
+					assert.False(t, seqs[seq-1], assertMsg)
+					seqs[seq-1] = true
+					err := state.ReleaseRequestSeq(uint64(seq))
+					assert.NoError(t, err, assertMsg)
+				} else {
+					assert.True(t, seqs[seq-1])
+				}
+			}
+		}()
+	}
+
+	wg.Wait()
+}
+
+func testPrepareRequestSeq(t *testing.T) {
+	s := New()
+
+	cases := []struct {
+		desc string
+		seq  int
+
+		release bool
+		prepare bool
+
+		ok bool
+	}{{
+		desc:    "Release and prepare first ID",
+		seq:     100,
+		release: true,
+		prepare: true,
+		ok:      true,
+	}, {
+		desc:    "Prepare the same ID",
+		seq:     100,
+		prepare: true,
+		ok:      false,
+	}, {
+		desc:    "Prepare older ID",
+		seq:     50,
+		prepare: true,
+		ok:      false,
+	}, {
+		desc:    "Prepare before release",
+		seq:     200,
+		prepare: true,
+		ok:      false,
+	}, {
+		desc:    "Release and prepare another ID",
+		seq:     200,
+		release: true,
+		prepare: true,
+		ok:      true,
+	}}
+
+	for _, c := range cases {
+		seq := uint64(c.seq)
+		if c.release {
+			new := s.CaptureRequestSeq(seq)
+			require.True(t, new, c.desc)
+
+			err := s.ReleaseRequestSeq(seq)
+			require.NoError(t, err, c.desc)
+		}
+		if c.prepare {
+			err := s.PrepareRequestSeq(seq)
+			if c.ok {
+				require.NoError(t, err, c.desc)
+			} else {
+				require.Error(t, err, c.desc)
+			}
+		}
+	}
+}
+
+func testRetireRequestSeq(t *testing.T) {
+	s := New()
+
+	cases := []struct {
+		desc string
+		seq  int
+
+		prepare bool
+		retire  bool
+
+		ok bool
+	}{{
+		desc:    "Prepare and retire first ID",
+		seq:     100,
+		prepare: true,
+		retire:  true,
+		ok:      true,
+	}, {
+		desc:   "Retire the same ID",
+		seq:    100,
+		retire: true,
+		ok:     false,
+	}, {
+		desc:   "Retire older ID",
+		seq:    50,
+		retire: true,
+		ok:     false,
+	}, {
+		desc:   "Retire before prepare",
+		seq:    200,
+		retire: true,
+		ok:     false,
+	}, {
+		desc:    "Prepare and retire another ID",
+		seq:     200,
+		prepare: true,
+		retire:  true,
+		ok:      true,
+	}}
+
+	for _, c := range cases {
+		seq := uint64(c.seq)
+		if c.prepare {
+			new := s.CaptureRequestSeq(seq)
+			require.True(t, new, c.desc)
+
+			err := s.ReleaseRequestSeq(seq)
+			require.NoError(t, err, c.desc)
+
+			err = s.PrepareRequestSeq(seq)
+			require.NoError(t, err, c.desc)
+		}
+		if c.retire {
+			err := s.RetireRequestSeq(seq)
+			if c.ok {
+				require.NoError(t, err, c.desc)
+			} else {
+				require.Error(t, err, c.desc)
+			}
+		}
+	}
+}
